Restructure `Package.swift` (#2462)
This might be a controversial PR! 🍿

# Motivation
It always bothered me how our `Package.swift` is structured. First, we had a huge list of `targets` at the top that distracted from the `products`. The `products` are the important part for our adopters since those are the actual modules they can use. Secondly, the formatting of the targets made it hard to follow and resulted in unnecessary merge conflicts.

# Modification
This PR moves the targets back to the bottom and splits all the various array elements into separate elements.

# Result
Easier to understand `Package.swift` and less merge conflicts.