Fix 5.9 and nightly allocation limits (#2485)
# Motivation
Our 5.9 and nightly CI has been failing for some time now due to flaky allocation tests.

# Modification
This PR slightly changes some shared infra for how we ran some of our allocation tests. We are now waiting for the client and server channel to close so that allocations are more stable.

# Results
Green CI again