Fix flaky `testRemovesAllHTTPRelatedHandlersAfterUpgrade` test
# Motivation
This test has been flaky since we only waited for the untyped handler to be removed but we also execute this test for the typed test subclass.

# Modification
This PR fixes the test by also waiting for the typed handler to no longer be present in the pipeline.

# Result
Less flaky tests.