Implement single-use resolvers wrapper
### Motivation:

Closes https://github.com/apple/swift-nio/issues/2434

### Modifications:
Implemented a new `DynamicResolver` class that wraps a use-exactly-once resolver and manages its lifecycle, creating a new instance of the resolver each time a new resolution is performed.

### Result:
It's now possible to wrap a single-use resolver in a `DynamicResolver` to use it in situations where it might be used never or more than once.