Support asynchronous MTELG construction
Motivation

Currently when a MultithreadedEventLoopGroup is constructed we
block the calling thread until all the background threads launch.
Generally this is ok (we don't expect it to take long), but it can
cause issues when the launching thread is high priority, as we will
wait for those low priority threads to get time to execute.

Additionally, this causes nuisance warnings for users: see
https://github.com/apple/swift-nio/issues/2223 for more.

To make this safe, we need to make it possible to construct a
SelectableEventLoop instance before the owning thread has launched.
This requires us to make the existing strong stored `thread`
property into something that can be lazily initialized.

We need this property in order to correctly implement `inEventLoop`.
The good news is that so long as the first thing the background thread
does is initialize the property, we can draw a confident signal from
this value being `nil`: we know that we aren't on the EL thread. After
all, if we were, we would have already initialized that field!

This calls for your friend and mine, `ManagedAtomicLazyReference`.

Modifications

- Move the stored Thread object for SelectableEventLoop and
    Selector into a `ManagedAtomicLazyReference`.
- Provide convenience accessors that force-unwrap the EL.
- Fix a few specific use-cases that may actually run before the EL
    is provided.
- Update a couple of tests that can observe this change.

Result

We can now construct MTLEG without waiting for the background
threads to spawn, and everything is fully functional.