Suppress the performance checker warning by being sneaky (#2620)
Motivation:

Users are receiving thread performance checker warnings when
they create MTELGs. This is largely unnecessary: MTELGs are
not typically created on the fly but instead once, at program
startup. That means the issue here isn't really all that
profound.

While we could work around it by substantially rearchitecting
the event loops (see #2618), an easier fix is to observe that
the thread performance checker doesn't fire when we use
condition variables.

Modifications:

Remove the use of the DispatchGroup and replace it with a
condition variable.

Result:

Simpler code and no thread performance checker warning.