Fix up the markdown output generated by `scripts/analyze_performance_results.rb` (#2651)
Motivation:

`scripts/analyze_performance_results.rb` has `-o markdown`. But the output that it currently generates isn't compliant. It outputs:
`|--|--|...` as the header separator. There need to be three or more `-` in that (or `:` replacing one of the dashes).

I've also put the test names in code blocks, this improves the rendering slightly.

Modifications:

* Fix header separator output in markdown
* Put the test names in code blocks

Result:

Markdown output renders properly in standard markdown viewers (github, MacDown).