Expose `NIOThreadPool.numberOfThreads` publicly. (#2676)
Motivation:

`numberOfThreads` is a required parameter to the public constructor, it's therefore known to the caller. Any use case that wants to optimise its logic for the parallelism offered by the particular `NIOThreadPool` needs to carry the number of threads separately from the `NIOThreadPool` itself. We should exposes `numberOfThreads` publicly from `NIOThreadPool` to avoid this complexity and possible bugs.

This is especially true for `NIOThreadPool.singleton` where the decision on the number of threads used is achieved through some complex and dynamic logic. It's therefore best not to second guess the result of those calculations.

Modifications:

Flip the visibility of `numberOfThreads` in `NIOThreadPool` to `public`.

Result:

Some complexity gets removed from client packages.

Co-authored-by: Cory Benfield <lukasa@apple.com>