Add a NIOFileSystem shim
Motivation:

Renaming the `NIOFileSystem` module to `_NIOFileSystem` is a breaking
change (which is okay, it's documented to be unstable API). However,
users will be presented with a wall of build errors on updating to the
next version of NIO. We can make the transition slightly easier for
them.

Modifications:

Add back a `NIOFileSystem` module which re-exports `_NIOFileSystem` and
add it as a target to the `_NIOFileSystem` product. This makes both
modules available to users. We can then remove `NIOFileSystem` after a
grace period.

Result:

Users can still build using `NIOFileSystem` but can transition to
`_NIOFileSystem`.