[GHA] Cxx interoperability compatibility check
# Motivation

Another reusable check is to make sure that all library products of a package are successfully building when consumed from a module that has Cxx interoperability enabled.

# Modification

This PR adds a new check to the reusable workflow to check for Cxx interoperability compatibility.

# Result

This should be one of the last reusable workflow checks.